Fixed bug for final exercise
Didn't realize that the await was calling back my main function with only
one argument, which was therefore creating incorrect catch responses
since my main function expects 2 arguments but we only promised one within
the try block. This was therefore printing incorrect information within
the console log because we were checking names against the number within our
if condition. By making sure to provide the necessary expected arguments
for the callback function that we were fetching data from and resolving that
inconsistency, we clarified where each argument is pointing to so that main
function then has the correct data from the array to then check against.